+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+## [Unreleased]
+
 ## [4.1.10] - 2017-08-21
 
 * Use custom LoaderException for context loading errors.