+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+* Initial support for MariaDB SQL contexts.
+
 ## [4.1.10] - 2017-08-21
 
 * Use custom LoaderException for context loading errors.